Software/CANSAT_GUI/requirement.txt
pyqt5
pyserial
pyqtgraph
folium
